requirements.txt
ipython==2.1.0
numpy==1.8.1
scipy==0.14.0
matplotlib==1.3.1
clint==0.3.7
networkx==1.9
pyenchant==1.6.6
nltk==3.0.0
